moved bin_movie to io.binary, and moved io operations in detection module to main_datect() function
