+++ .dvc/.gitignore
@@ -60,3 +60,5 @@
 /tmp
 /tmp
 /tmp
+/tmp
+/tmp
