+++ suite2p/detection/detect.py
@@ -4,6 +4,7 @@ from pathlib import Path
 from . import sourcery, sparsedetect, chan2detect
 from .stats import ROI
 from .masks import create_cell_masks, create_neuropil_masks, create_cell_pix
+from ..io.binary import bin_movie
 
 
 def main_detect(ops):
@@ -12,7 +13,25 @@ def main_detect(ops):
     else:
         d0 = ops['diameter']
         dy, dx = (d0, d0) if isinstance(d0, int) else d0
-    stats = select_rois(dy=dy, dx=dx, Ly=ops['Ly'], Lx=ops['Lx'], max_overlap=ops['max_overlap'], sparse_mode=ops['sparse_mode'], ops=ops)
+
+    t0 = time.time()
+    bin_size = int(max(1, ops['nframes'] // ops['nbinned'], np.round(ops['tau'] * ops['fs'])))
+    print('Binning movie in chunks of length %2.2d' % bin_size)
+    mov = bin_movie(
+        filename=ops['reg_file'],
+        Ly=ops['Ly'],
+        Lx=ops['Lx'],
+        n_frames=ops['nframes'],
+        bin_size=bin_size,
+        bad_frames=np.where(ops['badframes'])[0] if 'badframes' in ops else (),
+        y_range=ops['yrange'],
+        x_range=ops['xrange'],
+    )
+    ops['nbinned'] = mov.shape[0]
+    print('Binned movie [%d,%d,%d], %0.2f sec.' % (mov.shape[0], mov.shape[1], mov.shape[2], time.time() - t0))
+
+
+    stats = select_rois(mov=mov, dy=dy, dx=dx, Ly=ops['Ly'], Lx=ops['Lx'], max_overlap=ops['max_overlap'], sparse_mode=ops['sparse_mode'], ops=ops)
     # extract fluorescence and neuropil
     t0 = time.time()
     cell_pix = create_cell_pix(stats, Ly=ops['Ly'], Lx=ops['Lx'], allow_overlap=ops['allow_overlap'])
@@ -37,12 +56,13 @@ def main_detect(ops):
     return cell_pix, cell_masks, neuropil_masks, stats, ops
 
 
-def select_rois(dy: int, dx: int, Ly: int, Lx: int, max_overlap: float, sparse_mode: bool, ops):
+def select_rois(mov: np.ndarray, dy: int, dx: int, Ly: int, Lx: int, max_overlap: float, sparse_mode: bool, ops):
+
     t0 = time.time()
     if sparse_mode:
-        ops, stats = sparsedetect.sparsery(high_pass=int(ops['high_pass']), ops=ops)
+        ops, stats = sparsedetect.sparsery(rez=mov, high_pass=int(ops['high_pass']), ops=ops)
     else:
-        ops, stats = sourcery.sourcery(ops)
+        ops, stats = sourcery.sourcery(mov=mov, ops=ops)
     print('Found %d ROIs, %0.2f sec' % (len(stats), time.time() - t0))
 
     rois = [ROI(ypix=stat['ypix'], xpix=stat['xpix'], lam=stat['lam'], dx=dx, dy=dy) for stat in stats]
