+++ suite2p/detection/utils.py
@@ -4,7 +4,6 @@ import numpy as np
 from numpy.linalg import norm
 from scipy.ndimage import gaussian_filter
 
-from ..io.binary import BinaryFile
 
 def binned_mean(mov: np.ndarray, bin_size) -> np.ndarray:
     """Returns an array with the mean of each time bin (of size 'bin_size')."""
@@ -29,27 +28,6 @@ def reject_frames(mov: np.ndarray, bad_indices: Sequence[int], mov_indices: Opti
 def crop(mov: np.ndarray, y_range: Tuple[int, int], x_range: Tuple[int, int]) -> np.ndarray:
     """Returns cropped frames of 'mov' encompassed by y_range and x_range."""
     return mov[:, slice(*y_range), slice(*x_range)]
-
-
-def bin_movie(filename: str, Ly: int, Lx: int, bin_size: int, n_frames: int, x_range: Tuple[int, int] = (), y_range: Tuple[int, int] = (), bad_frames: Sequence[int] = ()):
-    """Returns binned movie [nbins x Ly x Lx] from filename that has bad frames rejected and cropped to (y_range, x_range)"""
-    with BinaryFile(Ly=Ly, Lx=Lx, read_file=filename) as f:
-        batches = []
-        batch_size = min(n_frames - len(bad_frames), 500) // bin_size * bin_size
-        for indices, data in f.iter_frames(batch_size=batch_size):
-
-            if len(x_range) and len(y_range):
-                data = crop(mov=data, y_range=y_range, x_range=x_range)
-
-            if len(bad_frames) > 0:
-                data = reject_frames(mov=data, bad_indices=bad_frames, mov_indices=indices, reject_threshold=0.5)
-
-            if data.shape[0] >= batch_size:  # todo: drops the end of the movie
-                dbin = binned_mean(mov=data, bin_size=bin_size)
-                batches.append(dbin)
-
-    mov = np.vstack(batches)
-    return mov
 
 
 def high_pass_gaussian_filter(mov: np.ndarray, width: int) -> np.ndarray:
